requirements.txt
pyTelegramBotAPI==4.15.0
aiogram==3.4.1
gspread==5.12.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
requests==2.31.0
